+++ python/aistore/sdk/obj/content_iterator/__init__.py
@@ -1,5 +1,5 @@
 #
-# Copyright (c) 2025, NVIDIA CORPORATION. All rights reserved.
+# Copyright (c) 2025-2026, NVIDIA CORPORATION. All rights reserved.
 #
 
 """
@@ -10,11 +10,13 @@ or in parallel using concurrent range-reads.
 """
 
 from aistore.sdk.obj.content_iterator.base import BaseContentIterProvider
+from aistore.sdk.obj.content_iterator.buffer import ParallelBuffer
 from aistore.sdk.obj.content_iterator.sequential import ContentIterProvider
 from aistore.sdk.obj.content_iterator.parallel import ParallelContentIterProvider
 
 __all__ = [
     "BaseContentIterProvider",
     "ContentIterProvider",
+    "ParallelBuffer",
     "ParallelContentIterProvider",
 ]